fix: address review notes on Pylon chat visibility
Clarify that re-binding onShow/onHide is intentional, share the test
Pylon mock, and open chat from the onboarding header through the store.
