+++ client/dashboard/src/hooks/usePylonChat.ts
@@ -8,8 +8,8 @@ import {
 } from "@/lib/pylon";
 
 /**
- * Shared Pylon chat visibility. `onShow` / `onHide` from the widget keep
- * every consumer (account menu, other Get Support buttons) in sync.
+ * Shared Pylon chat visibility. Widget `onShow` / `onHide` keep the
+ * account-menu label in sync with the chat window.
  */
 export function usePylonChat(): { isOpen: boolean; toggle: () => void } {
   useEffect(() => {
